Remove duplication moving ImageType to Jellyfin.Data
