+++ MediaBrowser.Controller/Dto/DtoOptions.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using MediaBrowser.Model.Entities;
+using Jellyfin.Data.Enums;
 using MediaBrowser.Model.Querying;
 
 namespace MediaBrowser.Controller.Dto
