+++ MediaBrowser.Controller/Entities/ItemImageInfo.cs
@@ -4,7 +4,7 @@
 
 using System;
 using System.Text.Json.Serialization;
-using MediaBrowser.Model.Entities;
+using Jellyfin.Data.Enums;
 
 namespace MediaBrowser.Controller.Entities
 {
