+++ MediaBrowser.Model/Dto/ImageInfo.cs
@@ -1,5 +1,5 @@
 #nullable disable
-using MediaBrowser.Model.Entities;
+using Jellyfin.Data.Enums;
 
 namespace MediaBrowser.Model.Dto
 {
