+++ MediaBrowser.Model/Dto/ImageOptions.cs
@@ -1,6 +1,6 @@
 #nullable disable
+using Jellyfin.Data.Enums;
 using MediaBrowser.Model.Drawing;
-using MediaBrowser.Model.Entities;
 
 namespace MediaBrowser.Model.Dto
 {
