+++ MediaBrowser.Model/Providers/RemoteImageInfo.cs
@@ -1,6 +1,6 @@
 #nullable disable
+using Jellyfin.Data.Enums;
 using MediaBrowser.Model.Dto;
-using MediaBrowser.Model.Entities;
 
 namespace MediaBrowser.Model.Providers
 {
